requirements.txt
jax==0.2.12
jaxlib==0.1.65
optax==0.0.6
dm-haiku==0.0.4